Move files according to specifications
