+++ frontend/src/features/setting-device/device-card.tsx
