+++ frontend/src/features/workflow/functions/delete-workflow.ts
@@ -1,4 +1,4 @@
-import { deleteWorkflowWithUI } from "../api/mocks/workflow-api";
+import { deleteWorkflowWithUI } from "../../api/mocks/workflow-api";
 
 export function deleteWorkflow(workflowID: number) {
   deleteWorkflowWithUI(workflowID);
